APSI-3 Poprawne handlowanie wyjątków.
