+++ backend/src/main/java/apsi/team3/backend/DTOs/ApiResponse.java
@@ -3,27 +3,27 @@ package apsi.team3.backend.DTOs;
 import java.util.*;
 
 public class ApiResponse<T> {
-    public T Data;
-    public Map<String, String> Errors;
-    public boolean Success;
+    public T data;
+    public Map<String, String> errors;
+    public boolean success;
 
     public ApiResponse(T data) {
-        Data = data;
-        Success = true;
-        Errors = null;
+        this.data = data;
+        success = true;
+        errors = null;
     }
 
     public ApiResponse(Map<String, String> errors) {
-        Errors = errors;
-        Success = false;
-        Data = null;
+        this.errors = errors;
+        success = false;
+        data = null;
     }
 
     public ApiResponse(String error, String key) {
-        Errors = new HashMap<String, String>() {{ 
+        errors = new HashMap<String, String>() {{ 
             put(key, error);
         }};
-        Success = false;
-        Data = null;
+        success = false;
+        data = null;
     }
 }
